+++ src/transformers/models/granite/modeling_granite/attention/__init__.py
@@ -1,5 +1,5 @@
 from ...configuration_granite import GraniteConfig
-from ..enums import AttentionHeadType, AttentionImplementation, PositionEmbeddingType
+from ...enums import AttentionHeadType, AttentionImplementation, PositionEmbeddingType
 from .base import Attention
 from .flash import FlashAttention
 from .math import MathAttention
