+++ src/transformers/models/granite/modeling_granite/attention/math.py
@@ -3,7 +3,7 @@ from typing import Tuple, Union
 import torch
 import torch.nn as nn
 
-from ..enums import AttentionHeadType, PositionEmbeddingType
+from ...enums import AttentionHeadType, PositionEmbeddingType
 from .base import Attention
