+++ src/transformers/models/granite/modeling_granite/layer.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 
 from .attention import get_attention_module
 from ..configuration_granite import GraniteConfig
-from .enums import AttentionHeadType, AttentionImplementation, NormalizationImplementation
+from ..enums import AttentionHeadType, AttentionImplementation, NormalizationImplementation
 from .mlp import MLP
 from .normalization import get_normalization_function
